+++ Content/Items/Melee/BurntSienna.cs
@@ -1,11 +1,11 @@
 ﻿using CalamityMod.Items;
 using CalamityMod.Projectiles.Healing;
+using CalamityWeaponRemake.Common;
+using Microsoft.Xna.Framework;
+using Terraria;
 using Terraria.DataStructures;
 using Terraria.ID;
 using Terraria.ModLoader;
-using Terraria;
-using Microsoft.Xna.Framework;
-using CalamityWeaponRemake.Common;
 
 namespace CalamityWeaponRemake.Content.Items.Melee
 {
@@ -54,7 +54,7 @@ namespace CalamityWeaponRemake.Content.Items.Melee
                 }
                 sengs = 0;
             }
-            
+
             return false;
         }
 
@@ -88,7 +88,7 @@ namespace CalamityWeaponRemake.Content.Items.Melee
         {
             if (Main.rand.NextBool(5))
             {
-                Dust.NewDust(new Vector2((float)hitbox.X, (float)hitbox.Y), hitbox.Width, hitbox.Height, 246);
+                Dust.NewDust(new Vector2(hitbox.X, hitbox.Y), hitbox.Width, hitbox.Height, 246);
             }
         }
     }
